+++ Modules/authenticateToken.js
@@ -1,0 +1,17 @@
+const jwt = require("jsonwebtoken");
+
+
+function authenticateToken(req, res, next) {
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    if (token == null) return res.sendStatus(401);
+  
+    jwt.verify(token, process.env.TOKEN, (err, user) => {
+      console.log(err);
+      if (err) return res.sendStatus(403);
+      req.user = user;
+      next();
+    });
+  }
+
+module.exports = { authenticateToken };
